doxygen comments added

+++ src/point.h
@@ -2,10 +2,61 @@
 #pragma once
 #include <stdlib.h>
 
+/**
+ * @brief Точка на двумерной плоскости
+ * 
+ */
 typedef struct Point2 point2_t;
 
+/**
+ * @brief Конструктор точки
+ * 
+ * @return point2_t* возвращает указатель на новую точку или NULL
+ * 
+ * @note возвращает NULL
+ */
 point2_t* pnew();
-void psetx(point2_t* p, float x);
-float pgetx(point2_t* p);
-void psety(point2_t* p, float y);
-float pgety(point2_t* p);
+
+/**
+ * @brief Сеттер абсциссы точки
+ * 
+ * @param p указатель на точку
+ * @param x новое значение абсциссы
+ * 
+ * @bug не обрабатывается указатель на NULL
+ */
+void psetx(point2_t* p, int x);
+
+/**
+ * @brief Геттер абсциссы точки
+ * 
+ * @param p указатель на точку
+ * @return int абсцисса точки
+ * 
+ * @bug не обрабатывается указатель на NULL
+ * 
+ */
+int pgetx(point2_t* p);
+
+
+/**
+ * @brief Сеттер ордианты точки
+ * 
+ * @param p указатель на точку
+ * @param x новое значение ордианты
+ * 
+ * @bug не обрабатывается указатель на NULL
+ * 
+ */
+void psety(point2_t* p, int y);
+
+/**
+ * @brief Геттер ординаты точки
+ * 
+ * @param p указатель на точку
+ * @return int ордината точки
+ * 
+ * @bug не обрабатывается указатель на NULL
+ * 
+ */
+int pgety(point2_t* p);
